Move and delete contents of MoveAfterRefactor
