+++ Client/src/Components/Shared/ResultTabs.scss
