+++ Client/src/Components/StepAnalysis/StepAnalysisExternalResult.tsx
