+++ Client/src/Components/StepAnalysis/StepAnalysisExternalResultFactory.tsx
