+++ Client/src/Components/StepAnalysis/StepAnalysisView.scss
