+++ Client/src/Components/StepAnalysis/images/analysis-background.png
